requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
bcrypt==4.0.1
Flask-Login==0.6.2
python-dotenv==1.0.0
ics==0.7.2
hugchat==0.1.0